requirements.txt
fastmcp
FastAPI
python-dotenv 

uvicorn
pydantic
requests
semantic-kernel

azure-monitor-opentelemetry
#azure-ai-evaluation
azure-ai-inference
azure-ai-projects
#opentelemetry-instrumentation-openai
# To have it in Telemetry in Azure AI Foundry
#opentelemetry-instrumentation-openai-v2
jsonref

gradio